refactor: extract babel setup helper into own module
This allows us to explicitly call it from other addons, like in `ember-decorators`, where we need to do that for backwards compatibility reasons.
